# Explore rearchitecting PenguinGraph's around closures.