--- a/Sources/PenguinGraphs/DepthFirstSearch.swift
+++ b/Sources/PenguinGraphs/DepthFirstSearch.swift
@@ -27,22 +27,41 @@ public enum VertexColor {
   case black
 }
 
-public enum DFSEvent<Graph: GraphProtocol> {
-  case start(Graph.VertexId)
-  case discover(Graph.VertexId)
-  case examine(Graph.EdgeId)
-  case treeEdge(Graph.EdgeId)
-  case backEdge(Graph.EdgeId)
-  case forwardOrCrossEdge(Graph.EdgeId)
-  case finish(Graph.VertexId)
+/// The events that occur during depth first search within a graph.
+///
+/// - SeeAlso: `IncidenceGraph.DFSCallback`
+public enum DFSEvent<SearchSpace: GraphProtocol> {
+  /// The start of the depth first search, recording the starting vertex.
+  case start(SearchSpace.VertexId)
+
+  /// When a new vertex is discovered in the search space.
+  case discover(SearchSpace.VertexId)
+
+  /// When an edge is traversed to explore the destination.
+  case examine(SearchSpace.EdgeId)
+
+  /// When the edge is determined to form part of the search tree.
+  ///
+  /// This event occurs when the destination of the edge is newly discovered.
+  case treeEdge(SearchSpace.EdgeId)
+
+  /// When the edge's destination is being processed (i.e. already on the stack).
+  case backEdge(SearchSpace.EdgeId)
+
+  /// When the edge's destination has already been processed.
+  case forwardOrCrossEdge(SearchSpace.EdgeId)
+
+  /// When all edges from a vertex have been explored.
+  case finish(SearchSpace.VertexId)
 }
 
 extension IncidenceGraph where Self: VertexListGraph {
 
+  /// A hook to observe events that occur during depth first search.
   public typealias DFSCallback = (DFSEvent<Self>, inout Self) throws -> Void
 
   /// Expores `self` depth-first starting at `source`, using `vertexVisitationState` to keep track of
-  /// visited vertices, invoking `visitor`'s methods to reflect changes to the search state.
+  /// visited vertices, invoking `callback` at key events of the search.
   ///
   /// - Note: this is a mutating method because the `vertexVisitationState` or `visitor` may store
   ///   data within the graph itself.
@@ -58,9 +77,12 @@ extension IncidenceGraph where Self: VertexListGraph {
     VertexVisitationState.Graph == Self,
     VertexVisitationState.Value == VertexColor
   {
+    assert(
+      vertexVisitationState.get(self, source) == .white,
+      "vertexVisitationState was not properly initialized.")
     try callback(.start(source), &self)
 
-    // We use an explicit stack to avoid a recursive implementation for performance.
+    // We use an explicit stack to avoid a recursive implementation for performance & scale.
     //
     // The stack contains the vertex we're traversing, as well as the (partially consumed) iterator
     // for the edges.
@@ -129,11 +151,10 @@ extension IncidenceGraph where Self: VertexListGraph {
 }
 
 extension IncidenceGraph where Self: VertexListGraph, VertexId: IdIndexable {
-  /// Expores `self` depth-first starting at `source`, invoking `visitor`'s methods to reflect
-  /// changes to the search state.
+  /// Expores `self` depth-first starting at `source`, invoking `callback` at key events during the
+  /// search.
   ///
-  /// - Note: this is a mutating method because the `visitor` may store data within the graph
-  ///   itself.
+  /// - Note: this is a mutating method because the `callback` may modify the graph.
   public mutating func depthFirstSearch(
     startingAt source: VertexId,
     callback: DFSCallback