Previously, `PenguinGraph` used Visitors to allow composition of
algorithms. This worked, but was somewhat unwieldly, and had a number
of drawbacks, including:

 - All graph traversals ended up being `throw`ing, which forced users to
   smear `try`'s in their code, because there was no way to express the
   `rethrows`-ingness within a protocol.
 - Because there's no `override` checking for implementing methods on
   protocols, it was very easy to implement a subtly wrong method
   signature. This would result in your method never being called.
 - I structured the BFS Visitor to itself contain the worklist within
   itself, so that Dijkstra's algorithm could modify it. This was a
   bit of a hack.

This new design, centered around closures, improves on all of these key
shortcomings of the previous design.

Performance benchmarks
======================

Before
------
```
name                                                              time            std         iterations
--------------------------------------------------------------------------------------------------------
AdjacencyList: build a complete graph of size 10                  47989.0 ns      ±  11.83 %  29416
AdjacencyList: build a complete graph of size 100                 4374958.0 ns    ±   7.11 %  304
AdjacencyList: build a complete graph of size 1000                435099068.0 ns  ±   2.91 %  3
AdjacencyList: DFS on oneEdge graph, counting visitor             1961937.5 ns    ±  14.90 %  696
AdjacencyList: BFS on oneEdge graph, counting visitor             1692833.0 ns    ±   7.64 %  871
AdjacencyList: Dijkstra on oneEdge graph, counting visitor        3508803.0 ns    ±   5.90 %  401
AdjacencyList: DFS on twoEdges graph, counting visitor            2565233.0 ns    ±   9.24 %  547
AdjacencyList: BFS on twoEdges graph, counting visitor            2314113.0 ns    ±   7.58 %  603
AdjacencyList: Dijkstra on twoEdges graph, counting visitor       5759319.0 ns    ±   6.72 %  258
AdjacencyList: DFS on completeGraph graph, counting visitor       579625778.5 ns  ±   0.80 %  2
AdjacencyList: BFS on completeGraph graph, counting visitor       603404920.5 ns  ±   2.51 %  2
AdjacencyList: Dijkstra on completeGraph graph, counting visitor  746910440.5 ns  ±   2.56 %  2
AdjacencyList: DFS on completeDAG graph, counting visitor         288798695.0 ns  ±   3.41 %  5
AdjacencyList: BFS on completeDAG graph, counting visitor         307119093.0 ns  ±   3.92 %  5
AdjacencyList: Dijkstra on completeDAG graph, counting visitor    289096968.0 ns  ±   2.94 %  5
AdjacencyList: Topological sort on completeDAG                    291341816.0 ns  ±   3.42 %  5
```

After
-----

```
name                                                              time            std         iterations
--------------------------------------------------------------------------------------------------------
AdjacencyList: build a complete graph of size 10                  48667.0 ns      ±  12.87 %  28688
AdjacencyList: build a complete graph of size 100                 4460344.0 ns    ±   7.61 %  326
AdjacencyList: build a complete graph of size 1000                449588758.0 ns  ±   2.48 %  3
AdjacencyList: DFS on oneEdge graph, counting visitor             1906472.0 ns    ±   8.44 %  727
AdjacencyList: BFS on oneEdge graph, counting visitor             1546362.0 ns    ±   8.32 %  915
AdjacencyList: Dijkstra on oneEdge graph, counting visitor        4079355.0 ns    ±  11.06 %  335
AdjacencyList: DFS on twoEdges graph, counting visitor            2691643.0 ns    ±   8.99 %  521
AdjacencyList: BFS on twoEdges graph, counting visitor            2097564.5 ns    ±   7.12 %  646
AdjacencyList: Dijkstra on twoEdges graph, counting visitor       5804325.5 ns    ±   8.40 %  240
AdjacencyList: DFS on completeGraph graph, counting visitor       597149572.5 ns  ±   1.36 %  2
AdjacencyList: BFS on completeGraph graph, counting visitor       646528025.5 ns  ±   0.66 %  2
AdjacencyList: Dijkstra on completeGraph graph, counting visitor  950868232.5 ns  ±   1.62 %  2
AdjacencyList: DFS on completeDAG graph, counting visitor         308286970.0 ns  ±   2.56 %  4
AdjacencyList: BFS on completeDAG graph, counting visitor         305959304.0 ns  ±   2.41 %  4
AdjacencyList: Dijkstra on completeDAG graph, counting visitor    404302281.0 ns  ±   2.67 %  3
AdjacencyList: Topological sort on completeDAG                    341198802.5 ns  ±   1.93 %  4
```

PAIR=@dabrahams