@saeta 2020-05-15
Running with cross module optimizations:

```
saeta-macbookpro2:penguin saeta$ time swift run -c release -Xswiftc -cross-module-optimization Benchmarks --filter "AdjacencyList"
[8/8] Linking Benchmarks
running AdjacencyList: build a complete graph of size 10... done! (2098.31 ms)
running AdjacencyList: build a complete graph of size 100... done! (1753.17 ms)
running AdjacencyList: build a complete graph of size 1000... done! (2229.53 ms)
running AdjacencyList: DFS on oneEdge graph, counting visitor... done! (1935.39 ms)
running AdjacencyList: BFS on oneEdge graph, counting visitor... done! (1570.37 ms)
running AdjacencyList: Dijkstra on oneEdge graph, counting visitor... done! (1467.91 ms)
running AdjacencyList: DFS on twoEdges graph, counting visitor... done! (1888.60 ms)
running AdjacencyList: BFS on twoEdges graph, counting visitor... done! (1414.69 ms)
running AdjacencyList: Dijkstra on twoEdges graph, counting visitor... done! (1665.24 ms)
running AdjacencyList: DFS on completeGraph graph, counting visitor... done! (1984.40 ms)
running AdjacencyList: BFS on completeGraph graph, counting visitor... done! (1853.46 ms)
running AdjacencyList: Dijkstra on completeGraph graph, counting visitor... done! (1866.28 ms)
running AdjacencyList: DFS on completeDAG graph, counting visitor... done! (1557.72 ms)
running AdjacencyList: BFS on completeDAG graph, counting visitor... done! (1522.68 ms)
running AdjacencyList: Dijkstra on completeDAG graph, counting visitor... done! (1496.76 ms)
running AdjacencyList: Topological sort on completeDAG... done! (1580.36 ms)

name                                                              time           std         iterations  
-------------------------------------------------------------------------------------------------------
AdjacencyList: build a complete graph of size 10                  5554.0 ns      ±  34.39 %  249576      
AdjacencyList: build a complete graph of size 100                 171107.5 ns    ±  22.58 %  9026        
AdjacencyList: build a complete graph of size 1000                6914846.5 ns   ±  14.53 %  200         
AdjacencyList: DFS on oneEdge graph, counting visitor             48011.0 ns     ±  21.12 %  26925       
AdjacencyList: BFS on oneEdge graph, counting visitor             120977.0 ns    ±  17.93 %  11454       
AdjacencyList: Dijkstra on oneEdge graph, counting visitor        1358657.0 ns   ±  14.60 %  938         
AdjacencyList: DFS on twoEdges graph, counting visitor            45674.0 ns     ±  13.32 %  28852       
AdjacencyList: BFS on twoEdges graph, counting visitor            118505.0 ns    ±  15.34 %  10396       
AdjacencyList: Dijkstra on twoEdges graph, counting visitor       2134791.0 ns   ±   9.61 %  652         
AdjacencyList: DFS on completeGraph graph, counting visitor       4471161.0 ns   ±  11.35 %  329         
AdjacencyList: BFS on completeGraph graph, counting visitor       3733026.0 ns   ±  10.44 %  379         
AdjacencyList: Dijkstra on completeGraph graph, counting visitor  41664100.0 ns  ±   5.50 %  34          
AdjacencyList: DFS on completeDAG graph, counting visitor         2139480.5 ns   ±  10.38 %  592         
AdjacencyList: BFS on completeDAG graph, counting visitor         1730760.0 ns   ±   8.03 %  751         
AdjacencyList: Dijkstra on completeDAG graph, counting visitor    18267805.0 ns  ±   4.34 %  69          
AdjacencyList: Topological sort on completeDAG                    1964697.5 ns   ±   7.75 %  678         
```