@shabalind 2020-05-15
@saeta Can you please provide updated before numbers with cross module optimisations? Otherwise it's hard to compare apples-to-apples.