@saeta 2020-05-15
Before performance with cross-module optimization:

```
saeta-macbookpro2:penguin saeta$ time swift run -c release -Xswiftc -cross-module-optimization Benchmarks --filter "AdjacencyList"
[5/5] Linking Benchmarks
running AdjacencyList: build a complete graph of size 10... done! (2127.44 ms)
running AdjacencyList: build a complete graph of size 100... done! (1599.99 ms)
running AdjacencyList: build a complete graph of size 1000... done! (2220.43 ms)
running AdjacencyList: DFS on oneEdge graph, counting visitor... done! (1929.94 ms)
running AdjacencyList: BFS on oneEdge graph, counting visitor... done! (1533.87 ms)
running AdjacencyList: Dijkstra on oneEdge graph, counting visitor... done! (1692.78 ms)
running AdjacencyList: DFS on twoEdges graph, counting visitor... done! (1972.04 ms)
running AdjacencyList: BFS on twoEdges graph, counting visitor... done! (1607.99 ms)
running AdjacencyList: Dijkstra on twoEdges graph, counting visitor... done! (1794.17 ms)
running AdjacencyList: DFS on completeGraph graph, counting visitor... done! (1966.21 ms)
running AdjacencyList: BFS on completeGraph graph, counting visitor... done! (1892.94 ms)
running AdjacencyList: Dijkstra on completeGraph graph, counting visitor... done! (1593.83 ms)
running AdjacencyList: DFS on completeDAG graph, counting visitor... done! (1658.09 ms)
running AdjacencyList: BFS on completeDAG graph, counting visitor... done! (1743.82 ms)
running AdjacencyList: Dijkstra on completeDAG graph, counting visitor... done! (1522.51 ms)
running AdjacencyList: Topological sort on completeDAG... done! (2121.70 ms)

name                                                              time            std         iterations  
--------------------------------------------------------------------------------------------------------
AdjacencyList: build a complete graph of size 10                  5839.0 ns       ±  30.13 %  245335      
AdjacencyList: build a complete graph of size 100                 159163.5 ns     ±   9.07 %  8906        
AdjacencyList: build a complete graph of size 1000                7106856.0 ns    ±   9.43 %  193         
AdjacencyList: DFS on oneEdge graph, counting visitor             43573.0 ns      ±  23.82 %  30262       
AdjacencyList: BFS on oneEdge graph, counting visitor             121771.0 ns     ±  12.10 %  11448       
AdjacencyList: Dijkstra on oneEdge graph, counting visitor        2561238.0 ns    ±  12.26 %  527         
AdjacencyList: DFS on twoEdges graph, counting visitor            45764.0 ns      ±  21.69 %  29197       
AdjacencyList: BFS on twoEdges graph, counting visitor            128229.5 ns     ±  13.70 %  11248       
AdjacencyList: Dijkstra on twoEdges graph, counting visitor       3817870.0 ns    ±  11.35 %  343         
AdjacencyList: DFS on completeGraph graph, counting visitor       4968374.0 ns    ±   6.66 %  281         
AdjacencyList: BFS on completeGraph graph, counting visitor       4551877.0 ns    ±   7.18 %  299         
AdjacencyList: Dijkstra on completeGraph graph, counting visitor  199448098.0 ns  ±   4.48 %  7           
AdjacencyList: DFS on completeDAG graph, counting visitor         2560007.0 ns    ±   7.13 %  531         
AdjacencyList: BFS on completeDAG graph, counting visitor         2409217.0 ns    ±  11.84 %  597         
AdjacencyList: Dijkstra on completeDAG graph, counting visitor    13304499.5 ns   ±   8.72 %  102         
AdjacencyList: Topological sort on completeDAG                    6581876.5 ns    ±   9.06 %  208         

real	0m37.559s
user	0m37.372s
sys	0m0.800s
```