@saeta 2020-05-15
While I haven't done any profiling, the numbers above indicate that this new closure-based design is generally slightly faster. Based on the design discussion earlier, I'll go ahead and polish this up and get this merged in.